Split TabMain into orchestrator + presentational TableView
Follow the SambaTabMain→ShareListView / NFSTabMain→NFSExportListView
convention. The new ConnectedClientsTableView owns the filter dropdown,
sortable headers, sort/filter state, and the table rows, taking
`clients` and `inFlight` as props and emitting `refresh` and
`kick(client)` events. ConnectedClientsTabMain keeps the manager
wiring, polling loop, in-flight gate, notification diff watch, and the
kick action — no behavior change.

ConnectedClientsTabMain drops 246 → 156 lines; ConnectedClientsTableView
is ~200 lines focused on presentation.
